arch: arm: Add mach-ipq support
These updates introduce support for mach-ipq SoCs
within the ARM architecture, including modifications
to the Makefile and Kconfig.

Additionally, they add support for the clock,
eMMC, UART, SMEM, and pinctrl drivers for the
IPQ5332, IPQ9574, and IPQ5424 SoCs.

The updates also include a list of bug fixes
for the 32-bit build in the serial_msm_geni driver
and clock-qcom.c.

Change-Id: I2e339faa57b1be578744c84af816bdf95d4f07eb
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>
Signed-off-by: Vandhiadevan Karunamoorthy <quic_vkarunam@quicinc.com>