mtd: spi-nor-core: Add fixups for s25fl129p flash
Fixups for the s25fl129p flash to support erase
handling when the device is configured with non-
uniform sector layout by default.

The fixup reads the CFR3V register to determine if
uniform sector mode is disabled. If so, it assigns a
custom erase hook to handle the 8 x 4KB sectors at
the bottom of the flash.

Change-Id: I2bd9ce9791acd13f2d9449ec2779559df26dbc06
Signed-off-by: Badhrinath S <badhrina@qti.qualcomm.com>