fs: fat: Perform sanity checks on getsize in get_fatent()
We do not perform a check on the value of getsize in get_fatent to
ensure that it will fit within the allocated buffer. For safety sake,
add a check now and if the value exceeds FATBUFBLOCKS use that value
instead. While not currently actively exploitable, it was in the past so
adding this check is worthwhile.

This addresses CVE-2025-24857 and was originally reported by Harvey
Phillips of Amazon Element55.

Change-Id: I65c3a144be103968c05edb68fa584ac5c0d80583
(cherry picked from commit 87d85139a96a39429120cca838e739408ef971a2)
Signed-off-by: Tom Rini <trini@konsulko.com>
Signed-off-by: Timple Raj M <timple@qti.qualcomm.com>
(cherry picked from commit 151277a513db38485c99311a7eebc5d1c90f2ada)